Fix int32 overflow in nopt_par_dgemm offsets
The per-thread A/C pointer offsets in the parallel dgemm wrapper were
computed in 32-bit int and overflowed once M*lda crossed 2^31 (large DF
Fock builds), giving a wild pointer and a crash. Widen to long. This is
a pre-existing bug in the original code.

+++ include/blas_link.h
@@ -61,9 +61,9 @@ inline void nopt_par_dgemm(const CBLAS_ORDER Layout, const CBLAS_TRANSPOSE Trans
         }
         cblas_dgemm(Layout, TransA, TransB, 
                     nopt_parallel_index[i+1]-nopt_parallel_index[i], N, K, alpha,
-                    A+nopt_parallel_index[i]*incr_a, lda, 
+                    A+(long)nopt_parallel_index[i]*incr_a, lda, 
                     B, ldb, beta,
-                    C+nopt_parallel_index[i]*ldc,ldc);
+                    C+(long)nopt_parallel_index[i]*ldc,ldc);
     }
 #ifdef _OPENBLAS
     openblas_set_num_threads(ntb);
